+++ gluecodium/src/test/resources/smoke/basic_types/output/cbridge/src/smoke/cbridge_BasicTypes.cpp
@@ -28,7 +28,7 @@ void smoke_BasicTypes_cache_swift_object_wrapper(_baseRef handle, const void* sw
     ::gluecodium::get_wrapper_cache().cache_wrapper(get_pointer<std::shared_ptr<::smoke::BasicTypes>>(handle)->get(), swift_pointer);
 }
 void smoke_BasicTypes_remove_swift_object_from_wrapper_cache(_baseRef handle) {
-    if (!wrapper_cache_is_alive) return;
+    if (!::gluecodium::WrapperCache::is_alive) return;
     ::gluecodium::get_wrapper_cache().remove_cached_wrapper(get_pointer<std::shared_ptr<::smoke::BasicTypes>>(handle)->get());
 }
 _baseRef smoke_BasicTypes_stringFunction(_baseRef input) {