+++ gluecodium/src/test/resources/smoke/errors/output/cbridge/src/smoke/cbridge_ErrorsInterface.cpp
@@ -30,7 +30,7 @@ void smoke_ErrorsInterface_cache_swift_object_wrapper(_baseRef handle, const voi
     ::gluecodium::get_wrapper_cache().cache_wrapper(get_pointer<std::shared_ptr<::smoke::ErrorsInterface>>(handle)->get(), swift_pointer);
 }
 void smoke_ErrorsInterface_remove_swift_object_from_wrapper_cache(_baseRef handle) {
-    if (!wrapper_cache_is_alive) return;
+    if (!::gluecodium::WrapperCache::is_alive) return;
     ::gluecodium::get_wrapper_cache().remove_cached_wrapper(get_pointer<std::shared_ptr<::smoke::ErrorsInterface>>(handle)->get());
 }
 extern "C" {