+++ gluecodium/src/test/resources/smoke/inheritance/output/cbridge/src/smoke/cbridge_ChildInterface.cpp
@@ -32,7 +32,7 @@ void smoke_ChildInterface_cache_swift_object_wrapper(_baseRef handle, const void
     ::gluecodium::get_wrapper_cache().cache_wrapper(get_pointer<std::shared_ptr<::smoke::ChildInterface>>(handle)->get(), swift_pointer);
 }
 void smoke_ChildInterface_remove_swift_object_from_wrapper_cache(_baseRef handle) {
-    if (!wrapper_cache_is_alive) return;
+    if (!::gluecodium::WrapperCache::is_alive) return;
     ::gluecodium::get_wrapper_cache().remove_cached_wrapper(get_pointer<std::shared_ptr<::smoke::ChildInterface>>(handle)->get());
 }
 extern "C" {