+++ gluecodium/src/test/resources/smoke/instances/output/cbridge/src/smoke/cbridge_SimpleClass.cpp
@@ -28,7 +28,7 @@ void smoke_SimpleClass_cache_swift_object_wrapper(_baseRef handle, const void* s
     ::gluecodium::get_wrapper_cache().cache_wrapper(get_pointer<std::shared_ptr<::smoke::SimpleClass>>(handle)->get(), swift_pointer);
 }
 void smoke_SimpleClass_remove_swift_object_from_wrapper_cache(_baseRef handle) {
-    if (!wrapper_cache_is_alive) return;
+    if (!::gluecodium::WrapperCache::is_alive) return;
     ::gluecodium::get_wrapper_cache().remove_cached_wrapper(get_pointer<std::shared_ptr<::smoke::SimpleClass>>(handle)->get());
 }
 _baseRef smoke_SimpleClass_getStringValue(_baseRef _instance) {