+++ gluecodium/src/test/resources/smoke/structs/output/cbridge/src/smoke/cbridge_Structs.cpp
@@ -33,7 +33,7 @@ void smoke_Structs_cache_swift_object_wrapper(_baseRef handle, const void* swift
     ::gluecodium::get_wrapper_cache().cache_wrapper(get_pointer<std::shared_ptr<::smoke::Structs>>(handle)->get(), swift_pointer);
 }
 void smoke_Structs_remove_swift_object_from_wrapper_cache(_baseRef handle) {
-    if (!wrapper_cache_is_alive) return;
+    if (!::gluecodium::WrapperCache::is_alive) return;
     ::gluecodium::get_wrapper_cache().remove_cached_wrapper(get_pointer<std::shared_ptr<::smoke::Structs>>(handle)->get());
 }
 _baseRef